Rename filter constants, consolidate reused function into filter libraries
