+++ ui/src/tests/LocalStorageFilters.test.tsx
@@ -22,7 +22,7 @@ import React from 'react';
 import {RenderResult} from '@testing-library/react';
 import {createBrowserHistory} from 'history';
 import {Router} from 'react-router-dom';
-import {LocalStorageFilters} from '../SortingAndFiltering/FilterConstants';
+import {LocalStorageFilters} from '../SortingAndFiltering/FilterLibraries';
 
 describe('Filter products', () => {
     class MockLocalStorage {
